Move some internal testing code into support directory and out of gem proper
